+++ feature/listing/src/main/kotlin/com/github/felipehjcosta/marvelapp/listing/presentation/CharacterItemViewModel.kt
@@ -1,3 +1,3 @@
 package com.github.felipehjcosta.marvelapp.listing.presentation
 
-class CharacterItemViewModel(val id: Int, val name: String, val image: String)
+class CharacterItemViewModel(val id: Long, val name: String, val image: String)
